Clean-up mongo code
 Mmove mongo wrappers to new file, re-structure functions order
