+++ pollsdata/mongo.go
@@ -17,7 +17,6 @@ package pollsdata
 import (
 	"context"
 	"errors"
-	"fmt"
 	"github.com/FabianWe/pollsweb"
 	"github.com/google/uuid"
 	"go.mongodb.org/mongo-driver/bson"
@@ -200,175 +199,6 @@ func NewMongoMeetingHandler(collection *mongo.Collection) *MongoMeetingHandler {
 	}
 }
 
-func (h *MongoMeetingHandler) InsertMeeting(ctx context.Context, meeting *MeetingModel) error {
-	_, insertErr := h.Collection.InsertOne(ctx, meeting)
-	return insertErr
-}
-
-func mongoDecodePollFromRaw(rawDocument bson.Raw) (AbstractPollModel, error) {
-	if validationErr := rawDocument.Validate(); validationErr != nil {
-		return nil, validationErr
-	}
-	pollType, lookupErr := rawDocument.LookupErr("type")
-	if lookupErr != nil {
-		return nil, lookupErr
-	}
-	pollTypeString, pollTypeStringOk := pollType.StringValueOK()
-	if !pollTypeStringOk {
-		return nil, errors.New("unable to decode poll type from bson: Not a string")
-	}
-	var res AbstractPollModel
-	switch pollTypeString {
-	case BasicPollStringName:
-		res = EmptyBasicPollModel()
-	case MedianPollStringName:
-		res = EmptyMedianPollModel()
-	case SchulzePollStringName:
-		res = EmptySchulzePollModel()
-	default:
-		return nil, fmt.Errorf("invalid poll type while parsing poll \"%s\"", pollTypeString)
-	}
-	if unmarshalErr := bson.Unmarshal(rawDocument, res); unmarshalErr != nil {
-		return nil, unmarshalErr
-	}
-	return res, nil
-}
-
-type mongoPollGroupModel struct {
-	*IdModel `bson:",inline"`
-	Name     string
-	Slug     string
-	Polls    []bson.Raw
-}
-
-func (m *mongoPollGroupModel) decodePolls() ([]AbstractPollModel, error) {
-	res := make([]AbstractPollModel, len(m.Polls))
-	for i, pollRaw := range m.Polls {
-		poll, pollErr := mongoDecodePollFromRaw(pollRaw)
-		if pollErr != nil {
-			return nil, fmt.Errorf("unable to decode poll (position %d): %w", i, pollErr)
-		}
-		res[i] = poll
-	}
-	return res, nil
-}
-
-func (m *mongoPollGroupModel) toPollGroupModel() (*PollGroupModel, error) {
-	polls, pollsErr := m.decodePolls()
-	if pollsErr != nil {
-		return nil, pollsErr
-	}
-	res := NewPollGroupModel(m.Name, m.Slug, polls)
-	// set the id, it's not provided in the constructor
-	res.IdModel = m.IdModel
-	return res, nil
-}
-
-type mongoMeetingModel struct {
-	*IdModel    `bson:",inline"`
-	Name        string
-	Slug        string
-	Created     time.Time
-	Period      string
-	MeetingTime time.Time
-	OnlineStart time.Time
-	OnlineEnd   time.Time
-	Voters      []*VoterModel
-	Groups      []*mongoPollGroupModel
-	LastUpdated time.Time
-	UpdateToken int64
-}
-
-func emptyMongoMeetingModel() *mongoMeetingModel {
-	return &mongoMeetingModel{
-		IdModel:     EmptyIdModel(),
-		Name:        "",
-		Slug:        "",
-		Created:     time.Time{},
-		Period:      "",
-		MeetingTime: time.Time{},
-		OnlineStart: time.Time{},
-		OnlineEnd:   time.Time{},
-		Voters:      nil,
-		Groups:      nil,
-		LastUpdated: time.Time{},
-		// no need to create a random here
-		UpdateToken: -1,
-	}
-}
-
-func (m *mongoMeetingModel) decodeGroups() ([]*PollGroupModel, error) {
-	res := make([]*PollGroupModel, len(m.Groups))
-	for i, internalGroup := range m.Groups {
-		groupModel, groupErr := internalGroup.toPollGroupModel()
-		if groupErr != nil {
-			return nil, fmt.Errorf("unable to decode group (position %d): %w", i, groupErr)
-		}
-		res[i] = groupModel
-	}
-	return res, nil
-}
-
-func (m *mongoMeetingModel) toMeetingModel() (*MeetingModel, error) {
-	groups, groupsErr := m.decodeGroups()
-	if groupsErr != nil {
-		return nil, groupsErr
-	}
-	// create new instance with the given values
-	res := NewMeetingModel(m.Name, m.Slug, m.Period, m.MeetingTime, m.OnlineStart, m.OnlineEnd,
-		m.Voters, groups)
-	// set the id (not provided in the constructor)
-	res.IdModel = m.IdModel
-	// also set last updated and update token
-	res.LastUpdated = m.LastUpdated
-	res.UpdateToken = m.UpdateToken
-	return res, nil
-}
-
-func (h *MongoMeetingHandler) getSingle(ctx context.Context, filter, key interface{}) (*MeetingModel, error) {
-	internalModel := emptyMongoMeetingModel()
-	err := h.Collection.FindOne(ctx, filter).Decode(internalModel)
-	if err != nil {
-		if errors.Is(err, mongo.ErrNoDocuments) {
-			return nil, NewEntryNotFoundError(meetingModelType, reflect.ValueOf(key), err)
-		}
-		return nil, err
-	}
-	return internalModel.toMeetingModel()
-}
-
-func (h *MongoMeetingHandler) generateFilter(args *MeetingQueryArgs) (bson.M, error) {
-	res := make(bson.M, 3)
-	if args.Id != nil {
-		res["_id"] = *args.Id
-	}
-	if args.Slug != nil {
-		res["slug"] = *args.Slug
-	}
-	if args.Name != nil {
-		res["name"] = *args.Slug
-	}
-	if len(res) == 0 {
-		return nil, ErrInvalidMeetingQuery
-	}
-	// check for optional args
-	if args.LastUpdated != nil {
-		res["lastupdated"] = *args.LastUpdated
-	}
-	if args.UpdateToken != nil {
-		res["updatetoken"] = *args.UpdateToken
-	}
-	return res, nil
-}
-
-func (h *MongoMeetingHandler) GetMeeting(ctx context.Context, args *MeetingQueryArgs) (*MeetingModel, error) {
-	filter, queryErr := h.generateFilter(args)
-	if queryErr != nil {
-		return nil, queryErr
-	}
-	return h.getSingle(ctx, filter, args)
-}
-
 func (h *MongoMeetingHandler) CreateIndexes(ctx context.Context) ([]string, error) {
 	indexes := []mongo.IndexModel{
 		h.nameIndex(),
@@ -464,4 +294,53 @@ func (h *MongoMeetingHandler) onlineVoteIndex() mongo.IndexModel {
 		},
 		Options: options.Index(),
 	}
+}
+
+func (h *MongoMeetingHandler) InsertMeeting(ctx context.Context, meeting *MeetingModel) error {
+	_, insertErr := h.Collection.InsertOne(ctx, meeting)
+	return insertErr
+}
+
+func (h *MongoMeetingHandler) getSingle(ctx context.Context, filter, key interface{}) (*MeetingModel, error) {
+	internalModel := emptyMongoMeetingModel()
+	err := h.Collection.FindOne(ctx, filter).Decode(internalModel)
+	if err != nil {
+		if errors.Is(err, mongo.ErrNoDocuments) {
+			return nil, NewEntryNotFoundError(meetingModelType, reflect.ValueOf(key), err)
+		}
+		return nil, err
+	}
+	return internalModel.toMeetingModel()
+}
+
+func (h *MongoMeetingHandler) generateFilter(args *MeetingQueryArgs) (bson.M, error) {
+	res := make(bson.M, 3)
+	if args.Id != nil {
+		res["_id"] = *args.Id
+	}
+	if args.Slug != nil {
+		res["slug"] = *args.Slug
+	}
+	if args.Name != nil {
+		res["name"] = *args.Slug
+	}
+	if len(res) == 0 {
+		return nil, ErrInvalidMeetingQuery
+	}
+	// check for optional args
+	if args.LastUpdated != nil {
+		res["lastupdated"] = *args.LastUpdated
+	}
+	if args.UpdateToken != nil {
+		res["updatetoken"] = *args.UpdateToken
+	}
+	return res, nil
+}
+
+func (h *MongoMeetingHandler) GetMeeting(ctx context.Context, args *MeetingQueryArgs) (*MeetingModel, error) {
+	filter, queryErr := h.generateFilter(args)
+	if queryErr != nil {
+		return nil, queryErr
+	}
+	return h.getSingle(ctx, filter, args)
 }
